^r Extract flag parsing and status from authpolicy manage.go, move-only (gofmt, vet, build, package tests pass; declaration set verified identical; internal restructure, no behavior change)
Step 2 of 3 for the authpolicy god-file split: the parse* flag handlers
move verbatim to manage_args.go and commandStatus plus its readiness
helpers to manage_status.go.
