+++ internal/authpolicy/manage.go
@@ -4,7 +4,6 @@
 package authpolicy
 
 import (
-	"flag"
 	"fmt"
 	"io"
 	"os"
@@ -14,7 +13,6 @@ import (
 
 	"github.com/omkhar/workcell/internal/authresolve"
 	"github.com/omkhar/workcell/internal/cliexit"
-	"github.com/omkhar/workcell/internal/pathutil"
 	"github.com/omkhar/workcell/internal/providerid"
 )
 
@@ -226,302 +224,6 @@ func usage(program string) string {
 // newCmdFlagSet constructs a flag.FlagSet whose error output and usage
 // callback are wired identically across every parse* helper below.
 // Each helper used to repeat this four-line preamble verbatim.
-func newCmdFlagSet(program, command string, stderr io.Writer) *flag.FlagSet {
-	fs := flag.NewFlagSet(command, flag.ContinueOnError)
-	fs.SetOutput(stderr)
-	fs.Usage = func() { fmt.Fprintln(stderr, usage(program)) }
-	return fs
-}
-
-// requireNonEmptyFlags emits the canonical "missing required flags"
-// usage+error when fs has positional args, or any of the provided
-// pointer values is the empty string.  Callers list the *string values
-// that the subcommand requires; the error is identical to what every
-// hand-rolled parser produced before, preserving the bash exit-code
-// contract.
-func requireNonEmptyFlags(fs *flag.FlagSet, required ...*string) error {
-	if fs.NArg() != 0 {
-		fs.Usage()
-		return fmt.Errorf("missing required flags")
-	}
-	for _, ptr := range required {
-		if ptr == nil || *ptr == "" {
-			fs.Usage()
-			return fmt.Errorf("missing required flags")
-		}
-	}
-	return nil
-}
-
-func parsePolicyPathArgs(program string, command string, args []string, stderr io.Writer) (string, error) {
-	fs := newCmdFlagSet(program, command, stderr)
-	policy := fs.String("policy", "", "")
-	if err := fs.Parse(args); err != nil {
-		return "", err
-	}
-	if err := requireNonEmptyFlags(fs, policy); err != nil {
-		return "", err
-	}
-	return resolveInputPath(*policy), nil
-}
-
-func parseInitArgs(program string, args []string, stderr io.Writer) (string, string, error) {
-	fs := newCmdFlagSet(program, "init", stderr)
-	policy := fs.String("policy", "", "")
-	managedRoot := fs.String("managed-root", "", "")
-	if err := fs.Parse(args); err != nil {
-		return "", "", err
-	}
-	// init uses a bespoke error string ("--policy and --managed-root are
-	// required") to match the bash predecessor's exact message; cannot
-	// share requireNonEmptyFlags's "missing required flags" wording.
-	if fs.NArg() != 0 || *policy == "" || *managedRoot == "" {
-		fs.Usage()
-		return "", "", fmt.Errorf("--policy and --managed-root are required")
-	}
-	return resolveInputPath(*policy), resolveInputPath(*managedRoot), nil
-}
-
-func parseSetArgs(program string, args []string, stderr io.Writer) (setOptions, error) {
-	fs := newCmdFlagSet(program, "set", stderr)
-	var opts setOptions
-	fs.StringVar(&opts.policyPath, "policy", "", "")
-	fs.StringVar(&opts.managedRoot, "managed-root", "", "")
-	fs.StringVar(&opts.agent, "agent", "", "")
-	fs.StringVar(&opts.credential, "credential", "", "")
-	fs.StringVar(&opts.sourceRaw, "source", "", "")
-	fs.StringVar(&opts.sourceBaseRaw, "source-base", "", "")
-	fs.StringVar(&opts.resolver, "resolver", "", "")
-	fs.BoolVar(&opts.ackHostResolver, "ack-host-resolver", false, "")
-	if err := fs.Parse(args); err != nil {
-		return setOptions{}, err
-	}
-	if err := requireNonEmptyFlags(fs, &opts.policyPath, &opts.managedRoot, &opts.agent, &opts.credential); err != nil {
-		return setOptions{}, err
-	}
-	if _, ok := SupportedAgents[opts.agent]; !ok {
-		return setOptions{}, fmt.Errorf("invalid agent: %s", opts.agent)
-	}
-	if _, ok := CredentialKeys[opts.credential]; !ok {
-		return setOptions{}, fmt.Errorf("invalid credential: %s", opts.credential)
-	}
-	opts.policyPath = resolveInputPath(opts.policyPath)
-	opts.managedRoot = resolveInputPath(opts.managedRoot)
-	return opts, nil
-}
-
-func parseUnsetArgs(program string, args []string, stderr io.Writer) (string, string, string, error) {
-	fs := newCmdFlagSet(program, "unset", stderr)
-	policy := fs.String("policy", "", "")
-	managedRoot := fs.String("managed-root", "", "")
-	credential := fs.String("credential", "", "")
-	if err := fs.Parse(args); err != nil {
-		return "", "", "", err
-	}
-	if err := requireNonEmptyFlags(fs, policy, managedRoot, credential); err != nil {
-		return "", "", "", err
-	}
-	if _, ok := CredentialKeys[*credential]; !ok {
-		return "", "", "", fmt.Errorf("invalid credential: %s", *credential)
-	}
-	return resolveInputPath(*policy), resolveInputPath(*managedRoot), *credential, nil
-}
-
-func parseStatusArgs(program string, args []string, stderr io.Writer) (statusOptions, error) {
-	fs := newCmdFlagSet(program, "status", stderr)
-	var opts statusOptions
-	fs.StringVar(&opts.policyPath, "policy", "", "")
-	fs.StringVar(&opts.agent, "agent", "", "")
-	fs.StringVar(&opts.mode, "mode", "strict", "")
-	if err := fs.Parse(args); err != nil {
-		return statusOptions{}, err
-	}
-	if err := requireNonEmptyFlags(fs, &opts.policyPath); err != nil {
-		return statusOptions{}, err
-	}
-	if opts.agent != "" {
-		if _, ok := SupportedAgents[opts.agent]; !ok {
-			return statusOptions{}, fmt.Errorf("invalid agent: %s", opts.agent)
-		}
-	}
-	if _, ok := SupportedModes[opts.mode]; !ok {
-		return statusOptions{}, fmt.Errorf("invalid mode: %s", opts.mode)
-	}
-	opts.policyPath = resolveInputPath(opts.policyPath)
-	return opts, nil
-}
-
-func parseWhyArgs(program string, args []string, stderr io.Writer) (whyOptions, error) {
-	fs := newCmdFlagSet(program, "why", stderr)
-	var opts whyOptions
-	fs.StringVar(&opts.policyPath, "policy", "", "")
-	fs.StringVar(&opts.agent, "agent", "", "")
-	fs.StringVar(&opts.mode, "mode", "", "")
-	fs.StringVar(&opts.credential, "credential", "", "")
-	if err := fs.Parse(args); err != nil {
-		return whyOptions{}, err
-	}
-	if err := requireNonEmptyFlags(fs, &opts.policyPath, &opts.agent, &opts.mode, &opts.credential); err != nil {
-		return whyOptions{}, err
-	}
-	if _, ok := SupportedAgents[opts.agent]; !ok {
-		return whyOptions{}, fmt.Errorf("invalid agent: %s", opts.agent)
-	}
-	if _, ok := SupportedModes[opts.mode]; !ok {
-		return whyOptions{}, fmt.Errorf("invalid mode: %s", opts.mode)
-	}
-	if _, ok := CredentialKeys[opts.credential]; !ok {
-		return whyOptions{}, fmt.Errorf("invalid credential: %s", opts.credential)
-	}
-	opts.policyPath = resolveInputPath(opts.policyPath)
-	return opts, nil
-}
-
-func parseBootstrapSummaryArgs(program string, args []string, stderr io.Writer) (bootstrapSummaryOptions, error) {
-	fs := newCmdFlagSet(program, "bootstrap-summary", stderr)
-	var opts bootstrapSummaryOptions
-	fs.StringVar(&opts.agent, "agent", "", "")
-	fs.StringVar(&opts.inputKinds, "input-kinds", "none", "")
-	fs.StringVar(&opts.resolvers, "resolvers", "none", "")
-	fs.StringVar(&opts.resolutionStates, "resolution-states", "none", "")
-	fs.StringVar(&opts.providerReadyStates, "provider-ready-states", "none", "")
-	if err := fs.Parse(args); err != nil {
-		return bootstrapSummaryOptions{}, err
-	}
-	if err := requireNonEmptyFlags(fs, &opts.agent); err != nil {
-		return bootstrapSummaryOptions{}, err
-	}
-	if _, ok := SupportedAgents[opts.agent]; !ok {
-		return bootstrapSummaryOptions{}, fmt.Errorf("invalid agent: %s", opts.agent)
-	}
-	return opts, nil
-}
-
-func resolveInputPath(raw string) string {
-	expanded, err := pathutil.ExpandUserPathStrictRequireNonEmpty(raw)
-	if err != nil {
-		return raw
-	}
-	if !filepath.IsAbs(expanded) {
-		expanded, err = filepath.Abs(expanded)
-		if err != nil {
-			return raw
-		}
-	}
-	return filepath.Clean(expanded)
-}
-
-func resolveOutputPath(raw string) string {
-	return resolveInputPath(raw)
-}
-
-func commandStatus(opts statusOptions, stdout io.Writer) error {
-	if _, err := os.Stat(opts.policyPath); os.IsNotExist(err) {
-		fmt.Fprintln(stdout, "injection_policy=none")
-		fmt.Fprintln(stdout, "default_injection_policy_path="+opts.policyPath)
-		fmt.Fprintln(stdout, "credential_keys=none")
-		fmt.Fprintln(stdout, "credential_input_kinds=none")
-		fmt.Fprintln(stdout, "credential_resolvers=none")
-		fmt.Fprintln(stdout, "credential_materialization=none")
-		fmt.Fprintln(stdout, "credential_resolution_states=none")
-		fmt.Fprintln(stdout, "provider_auth_ready_states=none")
-		fmt.Fprintln(stdout, "shared_auth_ready_states=none")
-		if opts.agent != "" {
-			fmt.Fprintln(stdout, "provider_auth_mode=none")
-			fmt.Fprintln(stdout, "provider_auth_modes=none")
-			fmt.Fprintln(stdout, "shared_auth_modes=none")
-			fmt.Fprintln(stdout, "github_auth_present=0")
-			printBootstrapSummary(stdout, defaultBootstrapSummary(opts.agent))
-		}
-		return nil
-	}
-
-	policy, policySources, err := loadPolicyBundle(opts.policyPath)
-	if err != nil {
-		return err
-	}
-	selected, err := selectedCredentials(policy, opts.agent, opts.mode)
-	if err != nil {
-		return err
-	}
-	for key, raw := range selected {
-		if err := validateStatusCredentialSource(key, raw, filepath.Dir(opts.policyPath)); err != nil {
-			return err
-		}
-	}
-	inputKinds := map[string]string{}
-	resolvers := map[string]string{}
-	materialization := map[string]string{}
-	resolutionStates := map[string]string{}
-	providerReadyStates := map[string]string{}
-	sharedReadyStates := map[string]string{}
-	for key, raw := range selected {
-		inputKinds[key] = credentialInputKind(raw)
-		if rawMap, ok := raw.(map[string]any); ok {
-			if resolver, ok := rawMap["resolver"].(string); ok {
-				resolvers[key] = resolver
-			}
-			if mat, ok := rawMap["materialization"].(string); ok {
-				materialization[key] = mat
-			}
-		}
-		if inputKinds[key] == "resolver" {
-			state, err := resolverReadinessForStatus(key, resolvers[key])
-			if err != nil {
-				return err
-			}
-			resolutionStates[key] = state
-		} else {
-			resolutionStates[key] = "source"
-		}
-	}
-	if opts.agent != "" {
-		providerReadyStates, sharedReadyStates, err = explainReadyStates(policy, filepath.Dir(opts.policyPath), opts.agent, opts.mode)
-		if err != nil {
-			return err
-		}
-	} else {
-		for key := range selected {
-			readyState := "ready"
-			if resolutionStates[key] == "configured-only" {
-				readyState = "configured-only"
-			}
-			if _, ok := SharedCredentialKeys[key]; ok {
-				sharedReadyStates[key] = readyState
-			} else {
-				providerReadyStates[key] = readyState
-			}
-		}
-	}
-
-	fmt.Fprintln(stdout, "policy_source_sha256="+compositePolicySHA256(policySources))
-	fmt.Fprintln(stdout, "credential_keys="+renderModes(sortedKeys(selected)))
-	fmt.Fprintln(stdout, "credential_input_kinds="+renderMap(inputKinds))
-	fmt.Fprintln(stdout, "credential_resolvers="+renderMap(resolvers))
-	fmt.Fprintln(stdout, "credential_materialization="+renderMap(materialization))
-	fmt.Fprintln(stdout, "credential_resolution_states="+renderMap(resolutionStates))
-	fmt.Fprintln(stdout, "provider_auth_ready_states="+renderMap(providerReadyStates))
-	fmt.Fprintln(stdout, "shared_auth_ready_states="+renderMap(sharedReadyStates))
-	if opts.agent != "" {
-		providerAuthModes := providerAuthModesForStatus(opts.agent, selected, resolutionStates)
-		sharedAuthModes := sharedAuthModesForStatus(selected, resolutionStates)
-		providerAuthMode := "none"
-		if len(providerAuthModes) > 0 {
-			providerAuthMode = providerAuthModes[0]
-		}
-		fmt.Fprintln(stdout, "provider_auth_mode="+providerAuthMode)
-		fmt.Fprintln(stdout, "provider_auth_modes="+renderModes(providerAuthModes))
-		fmt.Fprintln(stdout, "shared_auth_modes="+renderModes(sharedAuthModes))
-		if len(sharedAuthModes) > 0 {
-			fmt.Fprintln(stdout, "github_auth_present=1")
-		} else {
-			fmt.Fprintln(stdout, "github_auth_present=0")
-		}
-		printBootstrapSummary(stdout, summarizeBootstrap(opts.agent, selected, inputKinds, resolvers, resolutionStates, providerReadyStates))
-	}
-	return nil
-}
-
 func commandShow(policyPath string, stdout io.Writer) error {
 	policy, _, err := loadPolicyBundle(policyPath)
 	if err != nil {
@@ -762,89 +464,6 @@ func explainCredentialSelection(policy map[string]any, policyBase string, creden
 	report.reason = strings.Join(reasons, "; ")
 	return report, nil
 }
-
-func explainReadyStates(policy map[string]any, policyBase string, agent string, mode string) (map[string]string, map[string]string, error) {
-	providerReadyStates := map[string]string{}
-	sharedReadyStates := map[string]string{}
-	credentials, _ := policy["credentials"].(map[string]any)
-	if credentials == nil {
-		return providerReadyStates, sharedReadyStates, nil
-	}
-
-	relevant := map[string]struct{}{}
-	for key := range allowedCredentialsForAgent(agent) {
-		relevant[key] = struct{}{}
-	}
-	keys := make([]string, 0, len(relevant))
-	for key := range relevant {
-		keys = append(keys, key)
-	}
-	slices.Sort(keys)
-
-	for _, key := range keys {
-		if _, ok := credentials[key]; !ok {
-			continue
-		}
-		report, err := explainCredentialSelection(policy, policyBase, key, agent, mode)
-		if err != nil {
-			return nil, nil, err
-		}
-		if report.readiness == "" || report.readiness == "absent" {
-			continue
-		}
-		readyState := report.readiness
-		if credentialStateIsReady(report.readiness) {
-			readyState = "ready"
-		}
-		if _, ok := SharedCredentialKeys[key]; ok {
-			sharedReadyStates[key] = readyState
-		} else {
-			providerReadyStates[key] = readyState
-		}
-	}
-	return providerReadyStates, sharedReadyStates, nil
-}
-
-func resolverReadinessForStatus(key, resolver string) (string, error) {
-	if resolver == "" {
-		return "configured-only", nil
-	}
-	return authresolve.ProbeResolverReadiness(key, resolver)
-}
-
-func providerAuthModesForStatus(agent string, selected map[string]any, resolutionStates map[string]string) []string {
-	providerAuthModes := make([]string, 0)
-	for _, key := range statusOrder[agent] {
-		if _, ok := selected[key]; ok && credentialStateIsReady(resolutionStates[key]) {
-			providerAuthModes = append(providerAuthModes, key)
-		}
-	}
-	return providerAuthModes
-}
-
-func sharedAuthModesForStatus(selected map[string]any, resolutionStates map[string]string) []string {
-	sharedAuthModes := make([]string, 0)
-	for _, key := range []string{"github_hosts", "github_config"} {
-		if _, ok := selected[key]; ok && credentialStateIsReady(resolutionStates[key]) {
-			sharedAuthModes = append(sharedAuthModes, key)
-		}
-	}
-	return sharedAuthModes
-}
-
-func credentialStateIsReady(state string) bool {
-	switch state {
-	case "source", "resolved", "host-source", "ready":
-		return true
-	default:
-		return false
-	}
-}
-
-// Bootstrap-summary helpers (summarizeBootstrap, defaultBootstrapSummary,
-// bootstrapSummaryForCredential, bootstrapHandoffForReadiness,
-// bootstrapNextStepForReadiness, printBootstrapSummary,
-// printCredentialBootstrapSummary) live in bootstrap.go.
 
 func boolToInt(value bool) int {
 	if value {
